+++ modules/impls_shared/vertex_tools.cpp
@@ -23,12 +23,12 @@ __shared_func__ vertex get_barycentric_coords(matrix_coord &m_coords, vertex &v1
     return baryc;
 }
 
-__shared_func__ double length(vertex &vec)
+__shared_func__ float length(vertex &vec)
 {
     return sqrt(vec.x*vec.x + vec.y*vec.y + vec.z*vec.z);
 }
 
-__shared_func__ double dot(vertex &vec1, vertex &vec2)
+__shared_func__ float dot(vertex &vec1, vertex &vec2)
 {
     return
         vec1.x * vec2.x +
